+++ lib/features/onboarding/presentation/dashboard_guide_scope.dart
@@ -5,10 +5,19 @@ const dashboardShowcaseScope = 'dashboard_guide';
 class DashboardGuideKeys {
   final navigation = GlobalKey();
   final search = GlobalKey();
+  final entityTypeSelector = GlobalKey();
+  final filters = GlobalKey();
   final createItem = GlobalKey();
   final syncStatus = GlobalKey();
 
-  List<GlobalKey> get sequence => [navigation, search, createItem, syncStatus];
+  List<GlobalKey> get sequence => [
+    navigation,
+    search,
+    entityTypeSelector,
+    filters,
+    createItem,
+    syncStatus,
+  ];
 }
 
 class DashboardGuideScope extends InheritedWidget {